chore: migrate to vite static build
