+++ package.json
@@ -3,30 +3,35 @@
   "private": true,
   "version": "1.0.0",
   "scripts": {
-    "dev": "nuxi dev",
-    "build": "nuxi generate",
-    "start": "nuxi start"
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@nuxtjs/tailwindcss": "^4.2.1",
+    "@vueuse/head": "^0.7.4",
     "jszip": "^3.6.0",
-    "nuxt-edge": "^2.16.0-27322574.c1d02ea3"
+    "vue": "^3.2.26",
+    "vue-router": "^4.0.12"
   },
   "devDependencies": {
-    "@nuxt/bridge": "npm:@nuxt/bridge-edge",
-    "@nuxt/types": "^2.15.8",
-    "@nuxtjs/eslint-config-typescript": "^8.0.0",
-    "@nuxtjs/pwa": "^3.3.5",
-    "@nuxtjs/svg": "^0.3.0",
     "@types/node": "^14.14.37",
     "@types/pica": "^5.1.2",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
-    "eslint": "^7.23.0",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "@vitejs/plugin-vue": "^2.0.0",
+    "@vue/eslint-config-typescript": "^9.1.0",
+    "@vue/test-utils": "^2.0.0-rc.17",
+    "@vue/vue3-jest": "^27.0.0-alpha.4",
+    "eslint": "^8.4.1",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-vue": "^7.8.0",
-    "nuxt-windicss": "^2.1.1",
-    "typescript": "^4.5.3"
+    "eslint-plugin-vue": "^8.2.0",
+    "jest": "^27.4.5",
+    "typescript": "^4.5.3",
+    "vite": "^2.7.2",
+    "vite-plugin-pwa": "^0.11.11",
+    "vite-plugin-windicss": "^1.5.4",
+    "vite-svg-loader": "^3.1.1",
+    "vue-tsc": "^0.29.8"
   },
   "author": {
     "name": "Cristopher Namchee",
