+++ src/router/pages/about.vue
@@ -1,0 +1,246 @@
+<template>
+  <PageLayout>
+    <article class="w-full max-w-2xl mx-auto prose-lg lg:prose-xl py-8">
+      <h1 class="font-bold tracking-tight">
+        About Fav
+      </h1>
+
+      <section class="text-gray-500">
+        <h2 class="font-bold text-gray-700">
+          What is Fav?
+        </h2>
+
+        <p>
+          Fav is a next-generation icon generator for 2021 websites.
+          Fav allows you to generate favicon sets for modern browsers
+          from a single source file, eliminating you from a hassle of
+          manually generating multitudes different format of the said icon.
+          Fav is directly inspired by <a
+            class="link"
+            href="https://evilmartians.com/chronicles/how-to-favicon-in-2021-six-files-that-fit-most-needs"
+            target="_blank"
+            rel="noopener"
+            aria-label="EvilMartians Writeup"
+          >
+            this extensive writeup
+          </a>
+          written by <a
+            class="link"
+            href="https://evilmartians.com/"
+            target="_blank"
+            rel="noopener"
+            aria-label="EvilMartians Website"
+          >
+            Evil Martians.
+          </a>
+        </p>
+
+        <p>
+          Fav executes all the required processing steps directly
+          on your browser without any need to communicate with a backend server.
+          This means that fav <strong>
+            does not store any form of user-related data.
+          </strong>
+        </p>
+
+        <p>
+          The favicons used on this site are self-generated by fav itself.
+        </p>
+      </section>
+
+      <section class="text-gray-500">
+        <h2 class="font-bold text-gray-700">
+          Why another generator?
+        </h2>
+
+        <p>
+          Good question! While there are multitudes of favicons generator
+          out in the internet that you can already use today, they usually
+          aren't updated for modern use cases.
+        </p>
+
+        <p>
+          Back then, traditional browsers tend to have their own different
+          formats and standards to manage favicons, which should be adhered
+          to by responsible web developers. This situation is exacerbated by
+          the rise of smartphones which — unsurprisingly — have different
+          standards for managing favicons in mobile web browsers.
+        </p>
+
+        <p>
+          However, browsers are maturing at blazing speed everyday;
+          bringing more and more features on each updates,
+          including better favicon support. This eliminates the need for storing
+          20+ assets (which is generated by previous generators) on your
+          hosting just for showing a tiny image on a browser tab.
+        </p>
+      </section>
+
+      <section class="text-gray-500">
+        <h2 class="font-bold text-gray-700">
+          How to use Fav?
+        </h2>
+
+        <p>
+          Just head to the <router-link
+            class="link"
+            to="/"
+            aria-label="Fav Homepage"
+          >
+            homepage
+          </router-link> and follow the instructions there.
+          TL;DR, you'll have to perform the following steps:
+
+          <ol class="list">
+            <li>
+              Upload an image as a favicon base. Allowed formats are
+              <code>.png</code>, <code>.jpeg</code>, <code>.ico</code>,
+              and <code>.svg</code>. However,
+              <a
+                href="#caveats"
+                class="link"
+                aria-label="Anchor"
+              >
+                an SVG file is recommended.
+              </a>
+            </li>
+            <li>
+              Select your target platforms. Fav will only transform your asset
+              to be compatible with the selected platforms.
+            </li>
+            <li>
+              Generate your favicon set.
+            </li>
+          </ol>
+        </p>
+
+        <p>
+          After the favicon set is generated, extract the contents into the
+          <strong>public directory</strong> of
+          your website and include the icons in your HTML entry point.
+          For example:
+        </p>
+
+        <pre class="bg-gray-800 text-white overflow-x-auto">
+&lt;!-- start: head tag --&gt;
+  &lt;link rel="icon" href="/favicon.ico"&gt;
+  &lt;link rel="icon" href="/icon.svg" type="image/svg+xml"&gt;
+  &lt;link rel="apple-touch-icon" href="/apple-touch-icon.png"&gt;
+  &lt;link rel="manifest" href="/manifest.webmanifest"&gt;
+&lt;!-- end: head tag --&gt;
+        </pre>
+        <p>
+          If you're still confused about which line(s) should you include,
+          you can check the <strong>include HTML template</strong> checkbox
+          and Fav will generate a starter HTML template for you.
+        </p>
+      </section>
+
+      <section
+        id="caveats"
+        class="text-gray-500"
+      >
+        <h2 class="font-bold text-gray-700">
+          Caveats
+        </h2>
+
+        <p>
+          Fav is by no means perfect:
+        </p>
+
+        <ol class="list">
+          <li>
+            Since all process is done on your browser, the processing speed
+            is 100% dependant on your PC specs. But in general, it
+            should be fast.
+          </li>
+          <li>
+            Fav doesn't support
+            <a
+              href="https://en.wikipedia.org/wiki/Image_tracing"
+              target="_blank"
+              class="link"
+            >
+              image tracing
+            </a> <span class="italic">yet</span>.
+            Technically, transforming a raster to a vector is impossible.
+            To solve this problem, either provide an <code>.svg</code>
+            file as the base asset or uncheck the <strong>modern</strong>
+            platform.
+          </li>
+        </ol>
+      </section>
+
+      <section class="text-gray-500">
+        <h2 class="font-bold text-gray-700">
+          Contribution
+        </h2>
+
+        <p>
+          You can help improving Fav by visiting the
+          <a
+            href="https://github.com/Namchee/fav"
+            class="link"
+            rel="noopener"
+            target="_blank"
+            aria-label="Repository Link"
+          >
+            repository.
+          </a>
+          Any kind of contribution is welcome!
+        </p>
+      </section>
+    </article>
+  </PageLayout>
+</template>
+
+<script lang="ts">
+import { defineComponent } from 'vue';
+import PageLayout from '@/components/PageLayout.vue';
+import { useHead } from '@vueuse/head';
+
+export default defineComponent({
+  components: { PageLayout },
+
+  setup() {
+    useHead({
+      title: 'About Fav - Modern Favicon Generator',
+      meta: [
+        {
+          name: 'description',
+          content: 'Fav is a next-generation icon generator for 2021 websites.',
+        },
+        {
+          name: 'og:description',
+          content: 'Fav is a next-generation icon generator for 2021 websites.',
+        },
+        {
+          name: 'twitter:description',
+          content: 'Fav is a next-generation icon generator for 2021 websites.',
+        },
+      ],
+    });
+  },
+});
+</script>
+
+<style lang="postcss" scoped>
+.link {
+  @apply transition-colors text-indigo-500 hover:text-indigo-700;
+}
+
+.list {
+  list-style: none;
+  counter-reset: list-number;
+
+  & > li {
+    position: relative;
+    counter-increment: list-number;
+
+    &::before {
+      position: absolute;
+      content: counter(list-number) ". ";
+    }
+  }
+}
+</style>
