--- src/vue-shims.d.ts
@@ -1,5 +1,0 @@
-declare module '*.svg?inline' {
-  const content: any;
-
-  export default content;
-}
